Reorder params to evolve
Prior to this change, the xform-clauses of evolve were organised so that
the (optional) buffer preceeded the transducer.  This was so that the
transducer always appeared at the end of the arguments list.  It was not
the original intension to have the transducers output buffer preceed it
in the arguments and the documentation was poorly worded as a result.

This change reverses the ordering of the xform-clauses so that the
output buffer follows the transducer (if it appears at all).  The
documentation has been reworded (it may still read poorly, YMMV).